Move DebuggerVisualizerFile types from rustc_span to rustc_middle
